requirements.txt
openai
pandas
llama-index
jinja2
streamlit
spacy
langchain  # Adjust to the latest version available
# LlamaIndex  # Placeholder version, please check the correct one for LambIndex
mlflow
jupyterlab
jupyter
fastapi 
uvicorn 
pydantic
fastapi
uvicorn[standard]
python-dotenv

